@@ -5,7 +5,7 @@ class Program
 {
     static void Main()
     {
-        string connectionString = "Server=localhost;Database=master;Integrated Security=true;"
+        string connectionString = "Server=localhost;Database=master;Integrated Security=true;";
         // <Snippet1>
         // Assume that connectionString connects to master.  
         using (SqlConnection connection = new SqlConnection(connectionString))
@@ -13,7 +13,7 @@ class Program
         {
             connection.Open();
             command.Connection = connection;
-            command.Text = "USE DatabaseName";
+            command.CommandText = "USE DatabaseName";
             command.ExecuteNonQuery();
         }
             // </Snippet1>
