@@ -2,13 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#if NETFRAMEWORK
+#if NETFRAMEWORK && _WINDOWS
 
 using System;
 using System.Runtime.InteropServices;
 using System.Runtime.Versioning;
 using System.Security;
-using System.Text;
 
 namespace Interop.Windows.Kernel32
 {
@@ -21,18 +20,9 @@ namespace Interop.Windows.Kernel32
     {
         private const string Kernel32 = "kernel32.dll";
 
-        // http://msdn.microsoft.com/library/default.asp?url=/library/en-us/sysinfo/base/getcomputernameex.asp
-        [DllImport(Kernel32, CharSet = CharSet.Unicode, EntryPoint = "GetComputerNameExW", SetLastError = true)]
-        [ResourceExposure(ResourceScope.None)]
-        internal static extern int GetComputerNameEx(int nameType, StringBuilder nameBuffer, ref int bufferSize);
-
         [DllImport(Kernel32, CharSet = CharSet.Auto)]
         [ResourceExposure(ResourceScope.Process)]
         internal static extern int GetCurrentProcessId();
-
-        [DllImport(Kernel32, CharSet = CharSet.Auto, BestFitMapping = false, ThrowOnUnmappableChar = true)]
-        [ResourceExposure(ResourceScope.Process)]
-        internal static extern IntPtr GetModuleHandle([MarshalAs(UnmanagedType.LPTStr), In] string moduleName);
 
         [DllImport(Kernel32, CharSet = CharSet.Ansi, BestFitMapping = false, ThrowOnUnmappableChar = true, SetLastError = true)]
         [ResourceExposure(ResourceScope.None)]
