@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#if NET
+#if NET && _WINDOWS
 
 using System;
 using System.Runtime.InteropServices;
