@@ -4,6 +4,7 @@
 
 using System;
 using System.Collections;
+using System.Collections.Generic;
 using System.Data;
 using System.Data.Common;
 using System.Data.SqlTypes;
@@ -11,27 +12,26 @@ using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using System.Runtime.CompilerServices;
-
 using System.Security;
-using System.Security.Permissions;
+using System.Security.Authentication;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Transactions;
 using Microsoft.Data.Common.ConnectionString;
 using Microsoft.Data.SqlClient;
-using IsolationLevel = System.Data.IsolationLevel;
 using Microsoft.Identity.Client;
 using Microsoft.SqlServer.Server;
-using System.Security.Authentication;
-using System.Collections.Generic;
+using IsolationLevel = System.Data.IsolationLevel;
 
 #if NETFRAMEWORK
 using System.Reflection;
-using System.Runtime.ConstrainedExecution;
-using System.Runtime.InteropServices;
+using System.Security.Permissions;
+#endif
+
+#if _WINDOWS
 using System.Runtime.Versioning;
-using Interop.Windows.Kernel32;
+using Microsoft.Win32;
 #endif
 
 namespace Microsoft.Data.Common
@@ -45,6 +45,7 @@ namespace Microsoft.Data.Common
     /// This class is used so that there will be compile time checking of error messages.
     /// The resource Framework.txt will ensure proper string text based on the appropriate locale.
     /// </summary>
+    // @TODO: This file needs to be broken up
     internal static partial class ADP
     {
         // NOTE: Initializing a Task in SQL CLR requires the "UNSAFE" permission set (http://msdn.microsoft.com/en-us/library/ms172338.aspx)
@@ -76,6 +77,7 @@ namespace Microsoft.Data.Common
         /// </summary>
         internal static InvalidUdtException CreateInvalidUdtException(Type udtType, string resourceReasonName)
         {
+            // @TODO: Can we adopt the netcore version?
             InvalidUdtException e =
 #if NETFRAMEWORK
                 (InvalidUdtException)s_method.Invoke(null, new object[] { udtType, resourceReasonName });
@@ -361,13 +363,6 @@ namespace Microsoft.Data.Common
             return e;
         }
 
-        internal static ArgumentOutOfRangeException ArgumentOutOfRange(string message, string parameterName, object value)
-        {
-            ArgumentOutOfRangeException e = new(parameterName, value, message);
-            TraceExceptionAsReturnValue(e);
-            return e;
-        }
-
         internal static AuthenticationException SSLCertificateAuthenticationException(string message)
         {
             AuthenticationException e = new(message);
@@ -434,6 +429,46 @@ namespace Microsoft.Data.Common
             return InvalidEnumerationValue(typeof(CommandBehavior), (int)value);
         }
 
+
+        #if _UNIX
+        internal static object LocalMachineRegistryValue(string subkey, string queryvalue)
+        {
+            // No registry in non-Windows environments
+            return null;
+        }
+        #endif
+        #if _WINDOWS
+        [ResourceExposure(ResourceScope.Machine)]
+        [ResourceConsumption(ResourceScope.Machine)]
+        internal static object LocalMachineRegistryValue(string subkey, string queryvalue)
+        {
+            #if NETFRAMEWORK
+            new RegistryPermission(RegistryPermissionAccess.Read, $@"HKEY_LOCAL_MACHINE\{subkey}").Assert();
+            #endif
+
+            try
+            {
+                using (RegistryKey key = Registry.LocalMachine.OpenSubKey(subkey, false))
+                {
+                    return key?.GetValue(queryvalue);
+                }
+            }
+            catch (SecurityException e)
+            {
+                // Even though we assert permission - it's possible there are
+                // ACL's on registry that cause SecurityException to be thrown.
+                ADP.TraceExceptionWithoutRethrow(e);
+                return null;
+            }
+            #if NETFRAMEWORK
+            finally
+            {
+                CodeAccessPermission.RevertAssert();
+            }
+            #endif
+        }
+        #endif
+
         internal static void ValidateCommandBehavior(CommandBehavior value)
         {
             if (((int)value < 0) || (0x3F < (int)value))
@@ -456,9 +491,6 @@ namespace Microsoft.Data.Common
 #endif
             return InvalidEnumerationValue(typeof(Format), (int)value);
         }
-
-        internal static ArgumentOutOfRangeException NotSupportedUserDefinedTypeSerializationFormat(Format value, string method)
-            => NotSupportedEnumerationValue(typeof(Format), value.ToString(), method);
 
         internal static ArgumentException InvalidArgumentLength(string argumentName, int limit)
             => Argument(StringsHelper.GetString(Strings.ADP_InvalidArgumentLength, argumentName, limit));
@@ -954,6 +986,27 @@ namespace Microsoft.Data.Common
             };
         }
 
+        internal static ArgumentOutOfRangeException InvalidIsolationLevel(IsolationLevel value)
+        {
+            // @TODO: Use single line debug assert?
+            #if DEBUG
+            switch (value)
+            {
+                case IsolationLevel.Unspecified:
+                case IsolationLevel.Chaos:
+                case IsolationLevel.ReadUncommitted:
+                case IsolationLevel.ReadCommitted:
+                case IsolationLevel.RepeatableRead:
+                case IsolationLevel.Serializable:
+                case IsolationLevel.Snapshot:
+                    Debug.Fail("valid IsolationLevel " + value.ToString());
+                    break;
+            }
+            #endif
+
+            return InvalidEnumerationValue(typeof(IsolationLevel), (int)value);
+        }
+
         internal static InvalidOperationException NoConnectionString()
             => InvalidOperation(StringsHelper.GetString(Strings.ADP_NoConnectionString));
 
@@ -1078,9 +1131,6 @@ namespace Microsoft.Data.Common
 
         internal static bool CompareInsensitiveInvariant(string strvalue, string strconst)
             => 0 == CultureInfo.InvariantCulture.CompareInfo.Compare(strvalue, strconst, CompareOptions.IgnoreCase);
-
-        internal static int DstCompare(string strA, string strB) // this is null safe
-            => CultureInfo.CurrentCulture.CompareInfo.Compare(strA, strB, ADP.DefaultCompareOptions);
 
         internal static void SetCurrentTransaction(Transaction transaction) => Transaction.Current = transaction;
 
@@ -1237,11 +1287,6 @@ namespace Microsoft.Data.Common
         internal static Exception AmbiguousCollectionName(string collectionName)
             => Argument(StringsHelper.GetString(Strings.MDF_AmbiguousCollectionName, collectionName));
 
-        internal static Exception MissingDataSourceInformationColumn() => Argument(StringsHelper.GetString(Strings.MDF_MissingDataSourceInformationColumn));
-
-        internal static Exception IncorrectNumberOfDataSourceInformationRows()
-            => Argument(StringsHelper.GetString(Strings.MDF_IncorrectNumberOfDataSourceInformationRows));
-
         internal static Exception MissingRestrictionColumn() => Argument(StringsHelper.GetString(Strings.MDF_MissingRestrictionColumn));
 
         internal static Exception MissingRestrictionRow() => Argument(StringsHelper.GetString(Strings.MDF_MissingRestrictionRow));
@@ -1295,6 +1340,24 @@ namespace Microsoft.Data.Common
 
         internal static ArgumentException InvalidOffsetValue(int value)
             => Argument(StringsHelper.GetString(Strings.ADP_InvalidOffsetValue, value.ToString(CultureInfo.InvariantCulture)));
+
+        internal static ArgumentOutOfRangeException InvalidParameterDirection(ParameterDirection value)
+        {
+            // @TODO: Use single line debug assert?
+            #if DEBUG
+            switch (value)
+            {
+                case ParameterDirection.Input:
+                case ParameterDirection.Output:
+                case ParameterDirection.InputOutput:
+                case ParameterDirection.ReturnValue:
+                    Debug.Fail("valid ParameterDirection " + value.ToString());
+                    break;
+            }
+            #endif
+
+            return InvalidEnumerationValue(typeof(ParameterDirection), (int)value);
+        }
 
         internal static ArgumentException InvalidSizeValue(int value)
             => Argument(StringsHelper.GetString(Strings.ADP_InvalidSizeValue, value.ToString(CultureInfo.InvariantCulture)));
@@ -1396,249 +1459,5 @@ namespace Microsoft.Data.Common
         #endregion
 
         internal static readonly IntPtr s_ptrZero = IntPtr.Zero;
-#if NETFRAMEWORK
-#region netfx project only
-        internal static Task<T> CreatedTaskWithException<T>(Exception ex)
-        {
-            TaskCompletionSource<T> completion = new();
-            completion.SetException(ex);
-            return completion.Task;
-        }
-
-        //
-        // Helper Functions
-        //
-        internal static void CheckArgumentLength(string value, string parameterName)
-        {
-            CheckArgumentNull(value, parameterName);
-            if (0 == value.Length)
-            {
-                throw Argument(StringsHelper.GetString(Strings.ADP_EmptyString, parameterName)); // MDAC 94859
-            }
-        }
-
-        // IDbConnection.BeginTransaction, OleDbTransaction.Begin
-        internal static ArgumentOutOfRangeException InvalidIsolationLevel(IsolationLevel value)
-        {
-#if DEBUG
-            switch (value)
-            {
-                case IsolationLevel.Unspecified:
-                case IsolationLevel.Chaos:
-                case IsolationLevel.ReadUncommitted:
-                case IsolationLevel.ReadCommitted:
-                case IsolationLevel.RepeatableRead:
-                case IsolationLevel.Serializable:
-                case IsolationLevel.Snapshot:
-                    Debug.Assert(false, "valid IsolationLevel " + value.ToString());
-                    break;
-            }
-#endif
-            return InvalidEnumerationValue(typeof(IsolationLevel), (int)value);
-        }
-
-        // DBDataPermissionAttribute.KeyRestrictionBehavior
-        internal static ArgumentOutOfRangeException InvalidKeyRestrictionBehavior(KeyRestrictionBehavior value)
-        {
-#if DEBUG
-            switch (value)
-            {
-                case KeyRestrictionBehavior.PreventUsage:
-                case KeyRestrictionBehavior.AllowOnly:
-                    Debug.Assert(false, "valid KeyRestrictionBehavior " + value.ToString());
-                    break;
-            }
-#endif
-            return InvalidEnumerationValue(typeof(KeyRestrictionBehavior), (int)value);
-        }
-
-        // IDataParameter.Direction
-        internal static ArgumentOutOfRangeException InvalidParameterDirection(ParameterDirection value)
-        {
-#if DEBUG
-            switch (value)
-            {
-                case ParameterDirection.Input:
-                case ParameterDirection.Output:
-                case ParameterDirection.InputOutput:
-                case ParameterDirection.ReturnValue:
-                    Debug.Assert(false, "valid ParameterDirection " + value.ToString());
-                    break;
-            }
-#endif
-            return InvalidEnumerationValue(typeof(ParameterDirection), (int)value);
-        }
-
-        //
-        // DbConnectionOptions, DataAccess
-        //
-        internal static ArgumentException InvalidKeyname(string parameterName)
-        {
-            return Argument(StringsHelper.GetString(Strings.ADP_InvalidKey), parameterName);
-        }
-        internal static ArgumentException InvalidValue(string parameterName)
-        {
-            return Argument(StringsHelper.GetString(Strings.ADP_InvalidValue), parameterName);
-        }
-        internal static Exception InvalidMixedUsageOfAccessTokenAndCredential()
-        {
-            return ADP.InvalidOperation(StringsHelper.GetString(Strings.ADP_InvalidMixedUsageOfAccessTokenAndCredential));
-        }
-
-        //
-        // DBDataPermission, DataAccess, Odbc
-        //
-        internal static Exception InvalidXMLBadVersion()
-        {
-            return Argument(StringsHelper.GetString(Strings.ADP_InvalidXMLBadVersion));
-        }
-        internal static Exception NotAPermissionElement()
-        {
-            return Argument(StringsHelper.GetString(Strings.ADP_NotAPermissionElement));
-        }
-        internal static Exception PermissionTypeMismatch()
-        {
-            return Argument(StringsHelper.GetString(Strings.ADP_PermissionTypeMismatch));
-        }
-
-        //
-        // DbDataReader
-        //
-        internal static Exception NumericToDecimalOverflow()
-        {
-            return InvalidCast(StringsHelper.GetString(Strings.ADP_NumericToDecimalOverflow));
-        }
-
-        //
-        // DbDataAdapter
-        //
-        internal static InvalidOperationException ComputerNameEx(int lastError)
-        {
-            return InvalidOperation(StringsHelper.GetString(Strings.ADP_ComputerNameEx, lastError));
-        }
-
-        //
-        // SNI
-        //
-        internal static PlatformNotSupportedException SNIPlatformNotSupported(string platform) => new(StringsHelper.GetString(Strings.SNI_PlatformNotSupportedNetFx, platform));
-
-        // global constant strings
-        internal const float FailoverTimeoutStepForTnir = 0.125F; // Fraction of timeout to use in case of Transparent Network IP resolution.
-        internal const int MinimumTimeoutForTnirMs = 500; // The first login attempt in  Transparent network IP Resolution 
-
-        internal static readonly int s_ptrSize = IntPtr.Size;
-        internal static readonly IntPtr s_invalidPtr = new(-1); // use for INVALID_HANDLE
-
-        internal static readonly bool s_isWindowsNT = (PlatformID.Win32NT == Environment.OSVersion.Platform);
-        internal static readonly bool s_isPlatformNT5 = (ADP.s_isWindowsNT && (Environment.OSVersion.Version.Major >= 5));
-
-        [FileIOPermission(SecurityAction.Assert, AllFiles = FileIOPermissionAccess.PathDiscovery)]
-        [ResourceExposure(ResourceScope.Machine)]
-        [ResourceConsumption(ResourceScope.Machine)]
-        internal static string GetFullPath(string filename)
-        { // MDAC 77686
-            return Path.GetFullPath(filename);
-        }
-
-        // TODO: cache machine name and listen to longhorn event to reset it
-        internal static string GetComputerNameDnsFullyQualified()
-        {
-            const int ComputerNameDnsFullyQualified = 3; // winbase.h, enum COMPUTER_NAME_FORMAT
-            const int ERROR_MORE_DATA = 234; // winerror.h
-
-            string value;
-            if (s_isPlatformNT5)
-            {
-                int length = 0; // length parameter must be zero if buffer is null
-                                // query for the required length
-                                // VSTFDEVDIV 479551 - ensure that GetComputerNameEx does not fail with unexpected values and that the length is positive
-                int getComputerNameExError = 0;
-                if (0 == Kernel32Safe.GetComputerNameEx(ComputerNameDnsFullyQualified, null, ref length))
-                {
-                    getComputerNameExError = Marshal.GetLastWin32Error();
-                }
-                if ((getComputerNameExError != 0 && getComputerNameExError != ERROR_MORE_DATA) || length <= 0)
-                {
-                    throw ADP.ComputerNameEx(getComputerNameExError);
-                }
-
-                StringBuilder buffer = new(length);
-                length = buffer.Capacity;
-                if (0 == Kernel32Safe.GetComputerNameEx(ComputerNameDnsFullyQualified, buffer, ref length))
-                {
-                    throw ADP.ComputerNameEx(Marshal.GetLastWin32Error());
-                }
-
-                // Note: In Longhorn you'll be able to rename a machine without
-                // rebooting.  Therefore, don't cache this machine name.
-                value = buffer.ToString();
-            }
-            else
-            {
-                value = ADP.MachineName();
-            }
-            return value;
-        }
-
-        internal static IntPtr IntPtrOffset(IntPtr pbase, int offset)
-        {
-            if (4 == ADP.s_ptrSize)
-            {
-                return (IntPtr)checked(pbase.ToInt32() + offset);
-            }
-            Debug.Assert(8 == ADP.s_ptrSize, "8 != IntPtr.Size"); // MDAC 73747
-            return (IntPtr)checked(pbase.ToInt64() + offset);
-        }
-
-#endregion
-#else
-#region netcore project only
-
-        //
-        // COM+ exceptions
-        //
-        internal static PlatformNotSupportedException DbTypeNotSupported(string dbType) => new(StringsHelper.GetString(Strings.SQL_DbTypeNotSupportedOnThisPlatform, dbType));
-
-        // IDbConnection.BeginTransaction, OleDbTransaction.Begin
-        internal static ArgumentOutOfRangeException InvalidIsolationLevel(IsolationLevel value)
-        {
-#if DEBUG
-            switch (value)
-            {
-                case IsolationLevel.Unspecified:
-                case IsolationLevel.Chaos:
-                case IsolationLevel.ReadUncommitted:
-                case IsolationLevel.ReadCommitted:
-                case IsolationLevel.RepeatableRead:
-                case IsolationLevel.Serializable:
-                case IsolationLevel.Snapshot:
-                    Debug.Fail("valid IsolationLevel " + value.ToString());
-                    break;
-            }
-#endif
-            return InvalidEnumerationValue(typeof(IsolationLevel), (int)value);
-        }
-
-        // ConnectionUtil
-        internal static Exception IncorrectPhysicalConnectionType() => new ArgumentException(StringsHelper.GetString(StringsHelper.SNI_IncorrectPhysicalConnectionType));
-
-        // IDataParameter.Direction
-        internal static ArgumentOutOfRangeException InvalidParameterDirection(ParameterDirection value)
-        {
-#if DEBUG
-            switch (value)
-            {
-                case ParameterDirection.Input:
-                case ParameterDirection.Output:
-                case ParameterDirection.InputOutput:
-                case ParameterDirection.ReturnValue:
-                    Debug.Fail("valid ParameterDirection " + value.ToString());
-                    break;
-            }
-#endif
-            return InvalidEnumerationValue(typeof(ParameterDirection), (int)value);
-        }
-#endregion
-#endif
     }
 }
