@@ -69,13 +69,6 @@ namespace Microsoft.Data.ProviderBase
 
         private TransactionCompletedEventHandler _transactionCompletedEventHandler = null;
 
-        #if DEBUG
-        /// <summary>
-        /// Debug only counter to verify activate/deactivates are in sync.
-        /// </summary>
-        private int _activateCount;
-        #endif
-
         #endregion
 
         protected DbConnectionInternal() : this(ConnectionState.Open, true, false)
@@ -304,19 +297,6 @@ namespace Microsoft.Data.ProviderBase
         /// </summary>
         protected internal bool IsConnectionDoomed { get; private set; }
 
-        /// <summary>
-        /// Is this a connection that must be put in stasis (or is already in stasis) pending the
-        /// end of its transaction?
-        /// </summary>
-        /// <remarks>
-        /// If you want to have delegated transactions that are non-poolable, you had better
-        /// override this...
-        /// </remarks>
-        protected internal virtual bool IsNonPoolableTransactionRoot
-        {
-            get => false;
-        }
-
         /// <remarks>
         /// We use a weak reference to the owning object so we can identify when it has been
         /// garbage collected without throwing exceptions.
@@ -358,11 +338,6 @@ namespace Microsoft.Data.ProviderBase
             // Internal method called from the connection pooler so we don't expose
             // the Activate method publicly.
             SqlClientEventSource.Log.TryPoolerTraceEvent("<prov.DbConnectionInternal.ActivateConnection|RES|INFO|CPOOL> {0}, Activating", ObjectID);
-
-            #if DEBUG
-            int activateCount = Interlocked.Increment(ref _activateCount);
-            Debug.Assert(activateCount == 1, "activated multiple times?");
-            #endif
 
             Activate(transaction);
 
@@ -519,22 +494,6 @@ namespace Microsoft.Data.ProviderBase
             // Internal method called from the connection pooler so we don't expose
             // the Deactivate method publicly.
             SqlClientEventSource.Log.TryPoolerTraceEvent("<prov.DbConnectionInternal.DeactivateConnection|RES|INFO|CPOOL> {0}, Deactivating", ObjectID);
-
-            #if DEBUG
-            int origCount, newCount;
-            do
-            {
-                origCount = _activateCount;
-
-                if (origCount == 0)
-                {
-                  break;
-                }
-
-                newCount = origCount - 1;
-            }
-            while (Interlocked.CompareExchange(ref _activateCount, newCount, origCount) != origCount);
-            #endif
 
             SqlClientEventSource.Metrics.ExitActiveConnection();
 
@@ -716,9 +675,6 @@ namespace Microsoft.Data.ProviderBase
             Pool = connectionPool;
         }
 
-        internal void NotifyWeakReference(int message) =>
-            ReferenceCollection?.Notify(message);
-
         internal virtual void OpenConnection(DbConnection outerConnection, SqlConnectionFactory connectionFactory)
         {
             if (!TryOpenConnection(outerConnection, connectionFactory, null, null))
@@ -808,6 +764,13 @@ namespace Microsoft.Data.ProviderBase
         internal void RemoveWeakReference(object value) =>
             ReferenceCollection?.Remove(value);
 
+        /// <summary>
+        /// Idempotently resets the connection so that it may be recycled without leaking state.
+        /// May preserve transaction state if the connection is enlisted in a distributed transaction.
+        /// Should be called before the first action is taken on a recycled connection.
+        /// </summary>
+        internal abstract void ResetConnection();
+
         internal void SetInStasis()
         {
             IsTxRootWaitingForTxEnd = true;
@@ -845,6 +808,11 @@ namespace Microsoft.Data.ProviderBase
 
         #region Protected Methods
 
+        /// <summary>
+        /// Activates the connection, preparing it for active use.
+        /// An activated connection has an owner and is checked out from the connection pool (if pooling is enabled).
+        /// </summary>
+        /// <param name="transaction">The transaction in which the connection should enlist.</param>
         protected abstract void Activate(Transaction transaction);
 
         /// <summary>
@@ -861,6 +829,11 @@ namespace Microsoft.Data.ProviderBase
             throw ADP.InternalError(ADP.InternalErrorCode.AttemptingToConstructReferenceCollectionOnStaticObject);
         }
 
+        /// <summary>
+        /// Deactivates the connection, cleaning up any state as necessary.
+        /// A deactivated connection is one that is no longer in active use and does not have an owner.
+        /// A deactivated connection may be open (connected to a server) and is checked into the connection pool (if pooling is enabled).
+        /// </summary>
         protected abstract void Deactivate();
 
         protected internal void DoNotPoolThisConnection()
