@@ -8,6 +8,7 @@ namespace Microsoft.Data.SqlClient
 {
     internal static partial class LocalAppContextSwitches
     {
+        // @TODO: Replace with `bool?` since that's exactly how this is being used
         private enum Tristate : byte
         {
             NotInitialized = 0,
@@ -26,13 +27,17 @@ namespace Microsoft.Data.SqlClient
         private const string TruncateScaledDecimalString = @"Switch.Microsoft.Data.SqlClient.TruncateScaledDecimal";
         private const string IgnoreServerProvidedFailoverPartnerString = @"Switch.Microsoft.Data.SqlClient.IgnoreServerProvidedFailoverPartner";
         private const string EnableUserAgentString = @"Switch.Microsoft.Data.SqlClient.EnableUserAgent";
-#if NET
+
+        #if NET
         private const string GlobalizationInvariantModeString = @"System.Globalization.Invariant";
         private const string GlobalizationInvariantModeEnvironmentVariable = "DOTNET_SYSTEM_GLOBALIZATION_INVARIANT";
+
+        #if _WINDOWS
         private const string UseManagedNetworkingOnWindowsString = "Switch.Microsoft.Data.SqlClient.UseManagedNetworkingOnWindows";
-#else
+        #endif
+        #else
         private const string DisableTnirByDefaultString = @"Switch.Microsoft.Data.SqlClient.DisableTNIRByDefaultInConnectionString";
-#endif
+        #endif
 
         // this field is accessed through reflection in tests and should not be renamed or have the type changed without refactoring NullRow related tests
         private static Tristate s_legacyRowVersionNullBehavior;
@@ -47,28 +52,36 @@ namespace Microsoft.Data.SqlClient
         private static Tristate s_truncateScaledDecimal;
         private static Tristate s_ignoreServerProvidedFailoverPartner;
         private static Tristate s_enableUserAgent;
-#if NET
-        private static Tristate s_globalizationInvariantMode;
-        private static Tristate s_useManagedNetworking;
-#else
-        private static Tristate s_disableTnirByDefault;
-#endif
 
-#if NET
+        #if NET
+        private static Tristate s_globalizationInvariantMode;
+
+        #if _WINDOWS
+        private static Tristate s_useManagedNetworking;
+        #endif
+        #else
+        private static Tristate s_disableTnirByDefault;
+        #endif
+
+        #if NET
         static LocalAppContextSwitches()
         {
             IAppContextSwitchOverridesSection appContextSwitch = AppConfigManager.FetchConfigurationSection<AppContextSwitchOverridesSection>(AppContextSwitchOverridesSection.Name);
+
             try
             {
                 SqlAppContextSwitchManager.ApplyContextSwitches(appContextSwitch);
             }
             catch (Exception e)
             {
+                // @TODO: Adopt netcore style of trace logs
                 // Don't throw an exception for an invalid config file
                 SqlClientEventSource.Log.TryTraceEvent("<sc.{0}.ctor|INFO>: {1}", nameof(LocalAppContextSwitches), e);
             }
         }
-#endif
+        #endif
+
+        // @TODO: Sort by name
 
         /// <summary>
         /// In TdsParser, the ProcessSni function changed significantly when the packet
@@ -351,7 +364,8 @@ namespace Microsoft.Data.SqlClient
                 return s_enableUserAgent == Tristate.True;
             }
         }
-#if NET
+
+        #if NET
         /// <summary>
         /// .NET Core 2.0 and up supports Globalization Invariant mode, which reduces the size of the required libraries for
         /// applications which don't need globalization support. SqlClient requires those libraries for core functionality,
@@ -363,7 +377,7 @@ namespace Microsoft.Data.SqlClient
             {
                 if (s_globalizationInvariantMode == Tristate.NotInitialized)
                 {
-                    // Check if invariant mode is has been set by the AppContext switch directly
+                    // Check if invariant mode has been set by the AppContext switch directly
                     if (AppContext.TryGetSwitch(GlobalizationInvariantModeString, out bool returnedValue) && returnedValue)
                     {
                         s_globalizationInvariantMode = Tristate.True;
@@ -400,7 +414,19 @@ namespace Microsoft.Data.SqlClient
                 return s_globalizationInvariantMode == Tristate.True;
             }
         }
+        #else
+        /// <summary>
+        /// .NET Framework does not support Globalization Invariant mode, so this will always be false.
+        /// </summary>
+        public static bool GlobalizationInvariantMode
+        {
+            get => false;
+        }
+        #endif
 
+        #if NET
+
+        #if _WINDOWS
         /// <summary>
         /// When set to true, .NET Core will use the managed SNI implementation instead of the native SNI implementation.
         /// </summary>
@@ -437,30 +463,34 @@ namespace Microsoft.Data.SqlClient
                 return s_useManagedNetworking == Tristate.True;
             }
         }
-#else
+        #else
         /// <summary>
-        /// .NET Framework does not support Globalization Invariant mode, so this will always be false.
+        /// .NET Core on Unix does not support the native SNI, so this will always be true.
         /// </summary>
-        public const bool GlobalizationInvariantMode = false;
+        public static bool UseManagedNetworking => true;
+        #endif
 
+        #else
         /// <summary>
         /// .NET Framework does not support the managed SNI, so this will always be false.
         /// </summary>
-        public const bool UseManagedNetworking = false;
+        public static bool UseManagedNetworking => false;
+        #endif
 
+        #if NETFRAMEWORK
         /// <summary>
         /// Transparent Network IP Resolution (TNIR) is a revision of the existing MultiSubnetFailover feature.
         /// TNIR affects the connection sequence of the driver in the case where the first resolved IP of the hostname
         /// doesn't respond and there are multiple IPs associated with the hostname.
-        /// 
+        ///
         /// TNIR interacts with MultiSubnetFailover to provide the following three connection sequences:
         /// 0: One IP is attempted, followed by all IPs in parallel
         /// 1: All IPs are attempted in parallel
         /// 2: All IPs are attempted one after another
-        /// 
+        ///
         /// TransparentNetworkIPResolution is enabled by default. MultiSubnetFailover is disabled by default.
         /// To disable TNIR, you can enable the app context switch.
-        /// 
+        ///
         /// This app context switch defaults to 'false'.
         /// </summary>
         public static bool DisableTnirByDefault
@@ -481,6 +511,6 @@ namespace Microsoft.Data.SqlClient
                 return s_disableTnirByDefault == Tristate.True;
             }
         }
-#endif
+        #endif
     }
 }
