@@ -4,65 +4,51 @@
 
 #if NET
 
-using System.Threading;
+using System;
+
+#nullable enable
 
 namespace Microsoft.Data.SqlClient.ManagedSni
 {
     /// <summary>
     /// Global SNI settings and status
     /// </summary>
-    internal class SniLoadHandle
+    internal static class SniLoadHandle
     {
-        public static readonly SniLoadHandle SingletonInstance = new SniLoadHandle();
-
-        public readonly EncryptionOptions _encryptionOption = EncryptionOptions.OFF;
-        public ThreadLocal<SniError> _lastError = new ThreadLocal<SniError>(static () => new SniError(SniProviders.INVALID_PROV, 0, TdsEnums.SNI_SUCCESS, string.Empty));
-
-        private readonly uint _status = TdsEnums.SNI_SUCCESS;
+        [ThreadStatic]
+        private static SniError? s_lastError;
 
         /// <summary>
         /// Last SNI error
         /// </summary>
-        public SniError LastError
+        public static SniError LastError
         {
             get
             {
-                return _lastError.Value;
+                return s_lastError ??= SniError.Success;
             }
 
             set
             {
-                _lastError.Value = value;
+                s_lastError = value;
             }
         }
 
         /// <summary>
         /// SNI library status
         /// </summary>
-        public uint Status
-        {
-            get
-            {
-                return _status;
-            }
-        }
+        public const uint Status = TdsEnums.SNI_SUCCESS;
 
         /// <summary>
         /// Encryption options setting
         /// </summary>
-        public EncryptionOptions Options
-        {
-            get
-            {
-                return _encryptionOption;
-            }
-        }
+        public const EncryptionOptions Options = EncryptionOptions.OFF;
 
         /// <summary>
         /// Verify client encryption possibility
         /// </summary>
         // TODO: by adding support ENCRYPT_NOT_SUP, it could be calculated.
-        public bool ClientOSEncryptionSupport => true;
+        public const bool ClientOSEncryptionSupport = true;
     }
 }
 
