@@ -33,10 +33,16 @@ public sealed class LocalAppContextSwitchesHelper : IDisposable
     private readonly PropertyInfo _truncateScaledDecimalProperty;
     private readonly PropertyInfo _ignoreServerProvidedFailoverPartner;
     private readonly PropertyInfo _enableUserAgent;
-#if NET
+
+    #if NET
     private readonly PropertyInfo _globalizationInvariantModeProperty;
+    #endif
+    
+    #if NET && _WINDOWS
     private readonly PropertyInfo _useManagedNetworkingProperty;
-    #else
+    #endif
+    
+    #if NETFRAMEWORK
     private readonly PropertyInfo _disableTnirByDefaultProperty;
     #endif
 
@@ -63,12 +69,18 @@ public sealed class LocalAppContextSwitchesHelper : IDisposable
     private readonly Tristate _ignoreServerProvidedFailoverPartnerOriginal;
     private readonly FieldInfo _enableUserAgentField;
     private readonly Tristate _enableUserAgentOriginal;
-#if NET
+
+    #if NET
     private readonly FieldInfo _globalizationInvariantModeField;
     private readonly Tristate _globalizationInvariantModeOriginal;
+    #endif
+    
+    #if NET && _WINDOWS
     private readonly FieldInfo _useManagedNetworkingField;
     private readonly Tristate _useManagedNetworkingOriginal;
-    #else
+    #endif    
+
+    #if NETFRAMEWORK
     private readonly FieldInfo _disableTnirByDefaultField;
     private readonly Tristate _disableTnirByDefaultOriginal;
     #endif
@@ -169,15 +181,19 @@ public sealed class LocalAppContextSwitchesHelper : IDisposable
             "EnableUserAgent",
             out _enableUserAgent);
 
-#if NET
+        #if NET
         InitProperty(
             "GlobalizationInvariantMode",
             out _globalizationInvariantModeProperty);
-
+        #endif
+        
+        #if NET && _WINDOWS
         InitProperty(
             "UseManagedNetworking",
             out _useManagedNetworkingProperty);
-        #else
+        #endif
+
+        #if NETFRAMEWORK
         InitProperty(
             "DisableTnirByDefault",
             out _disableTnirByDefaultProperty);
@@ -253,17 +269,21 @@ public sealed class LocalAppContextSwitchesHelper : IDisposable
             out _enableUserAgentField,
             out _enableUserAgentOriginal);
 
-#if NET
+        #if NET
         InitField(
             "s_globalizationInvariantMode",
             out _globalizationInvariantModeField,
             out _globalizationInvariantModeOriginal);
-
+        #endif
+        
+        #if NET && _WINDOWS
         InitField(
             "s_useManagedNetworking",
             out _useManagedNetworkingField,
             out _useManagedNetworkingOriginal);
-        #else
+        #endif
+        
+        #if NETFRAMEWORK
         InitField(
             "s_disableTnirByDefault",
             out _disableTnirByDefaultField,
@@ -339,15 +359,19 @@ public sealed class LocalAppContextSwitchesHelper : IDisposable
             _enableUserAgentField,
             _enableUserAgentOriginal);
 
-#if NET
+        #if NET
         RestoreField(
             _globalizationInvariantModeField,
             _globalizationInvariantModeOriginal);
-
+        #endif
+        
+        #if NET && _WINDOWS
         RestoreField(
             _useManagedNetworkingField,
             _useManagedNetworkingOriginal);
-        #else
+        #endif
+        
+        #if NETFRAMEWORK
         RestoreField(
             _disableTnirByDefaultField,
             _disableTnirByDefaultOriginal);
@@ -450,7 +474,7 @@ public sealed class LocalAppContextSwitchesHelper : IDisposable
         get => (bool)_enableUserAgent.GetValue(null);
     }
 
-#if NET
+    #if NET
     /// <summary>
     /// Access the LocalAppContextSwitches.GlobalizationInvariantMode property.
     /// </summary>
@@ -458,7 +482,9 @@ public sealed class LocalAppContextSwitchesHelper : IDisposable
     {
         get => (bool)_globalizationInvariantModeProperty.GetValue(null);
     }
+    #endif
 
+    #if NET && _WINDOWS
     /// <summary>
     /// Access the LocalAppContextSwitches.UseManagedNetworking property.
     /// </summary>
@@ -466,7 +492,9 @@ public sealed class LocalAppContextSwitchesHelper : IDisposable
     {
         get => (bool)_useManagedNetworkingProperty.GetValue(null);
     }
-    #else
+    #endif
+    
+    #if NETFRAMEWORK
     /// <summary>
     /// Access the LocalAppContextSwitches.DisableTnirByDefault property.
     /// </summary>
@@ -580,7 +608,7 @@ public sealed class LocalAppContextSwitchesHelper : IDisposable
         set => SetValue(_enableUserAgentField, value);
     }
 
-#if NET
+    #if NET
     /// <summary>
     /// Get or set the LocalAppContextSwitches.GlobalizationInvariantMode switch value.
     /// </summary>
@@ -589,7 +617,9 @@ public sealed class LocalAppContextSwitchesHelper : IDisposable
         get => GetValue(_globalizationInvariantModeField);
         set => SetValue(_globalizationInvariantModeField, value);
     }
+    #endif
 
+    #if NET && _WINDOWS
     /// <summary>
     /// Get or set the LocalAppContextSwitches.UseManagedNetworking switch value.
     /// </summary>
@@ -598,7 +628,9 @@ public sealed class LocalAppContextSwitchesHelper : IDisposable
         get => GetValue(_useManagedNetworkingField);
         set => SetValue(_useManagedNetworkingField, value);
     }
-    #else
+    #endif
+    
+    #if NETFRAMEWORK
     /// <summary>
     /// Get or set the LocalAppContextSwitches.DisableTnirByDefault switch
     /// value.
