@@ -559,6 +559,10 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests.AlwaysEncrypted
                                 tryagain = true;
                                 break;
                             }
+                            else
+                            {
+                                throw;
+                            }
                         }
                         Assert.Equal(numberOfRows, rowsAffected);
                         tryagain = false;
@@ -2539,6 +2543,21 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests.AlwaysEncrypted
             Task readAsyncTask = ReadAsync(cmd, values, CommandBehavior.Default);
             readAsyncTask.GetAwaiter().GetResult();
 
+            // TODO(GH-3604): This section fails on Linux:
+            //
+            // Failed Microsoft.Data.SqlClient.ManualTesting.Tests.AlwaysEncrypted.ApiShould.TestRetryWhenAEEnclaveCacheIsStale(connectionString: "Data Source=tcp:10.0.0.4;Database=NORTHWIND;UID=sa"···) [45 s]
+            // EXEC : error Message:  [/mnt/vss/_work/1/s/build.proj]
+            //     Microsoft.Data.SqlClient.EnclaveDelegate+RetryableEnclaveQueryExecutionException : testing
+            //     Stack Trace:
+            //     at Microsoft.Data.SqlClient.SqlCommand.EndExecuteReaderAsync(IAsyncResult asyncResult) in /_/src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/SqlCommand.netcore.cs:line 1025
+            //     at Microsoft.Data.SqlClient.SqlCommand.<>c.<InternalExecuteReaderAsync>b__233_1(IAsyncResult asyncResult) in /_/src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/SqlCommand.netcore.cs:line 1569
+            //     at System.Threading.Tasks.TaskFactory`1.FromAsyncCoreLogic(IAsyncResult iar, Func`2 endFunction, Action`1 endAction, Task`1 promise, Boolean requiresSynchronization)
+            // --- End of stack trace from previous location ---
+            //     at Microsoft.Data.SqlClient.ManualTesting.Tests.AlwaysEncrypted.ApiShould.ReadAsync(SqlCommand sqlCommand, IList`1 values, CommandBehavior commandBehavior) in /_/src/Microsoft.Data.SqlClient/tests/ManualTests/AlwaysEncrypted/ApiShould.cs:line 2776
+            //     at Microsoft.Data.SqlClient.ManualTesting.Tests.AlwaysEncrypted.ApiShould.TestRetryWhenAEEnclaveCacheIsStale(String connectionString) in /_/src/Microsoft.Data.SqlClient/tests/ManualTests/AlwaysEncrypted/ApiShould.cs:line 2563
+            //     at System.RuntimeMethodHandle.InvokeMethod(Object target, Void** arguments, Signature sig, Boolean isConstructor)
+            //     at System.Reflection.MethodBaseInvoker.InvokeDirectByRefWithFewArgs(Object obj, Span`1 copyOfArgs, BindingFlags invokeAttr)
+            /*
 #if DEBUG
             CommandHelper.ForceThrowDuringGenerateEnclavePackage(cmd);
 
@@ -2558,7 +2577,8 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests.AlwaysEncrypted
             Task readAsyncTask2 = ReadAsync(cmd, values, CommandBehavior.Default);
             readAsyncTask2.GetAwaiter().GetResult();
 #endif
-
+            */
+            
             // revert the CEK change to the CustomerId column
             cmd.Parameters.Clear();
             cmd.CommandText = string.Format(alterCekQueryFormatString, _tableName, table.columnEncryptionKey1.Name);
@@ -3150,10 +3170,22 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests.AlwaysEncrypted
             }
             catch (AggregateException aggregateException)
             {
+                bool unexpected = false;
                 foreach (Exception ex in aggregateException.InnerExceptions)
                 {
-                    Assert.True(ex is SqlException, @"cancelling a command through cancellation token resulted in unexpected exception.");
-                    Assert.True(@"Operation cancelled by user." == ex.Message, @"cancelling a command through cancellation token resulted in unexpected error message.");
+                    if (ex is SqlException or InvalidOperationException)
+                    {
+                        Assert.Equal("Operation cancelled by user.", ex.Message);
+                    }
+                    else
+                    {
+                        unexpected = true;
+                        Console.WriteLine($"Cancellation produced non-SqlException: {ex}");
+                    }
+                }
+                if (unexpected)
+                {
+                    Assert.Fail("Unexpected exceptions encountered; see console for details.");
                 }
             }
 
